Rename a few modules
