+++ tlspuffin/src/experiment.rs
@@ -2,7 +2,7 @@ use std::{fmt::Display, fs::File, io, io::Write, path::Path, process::Command};
 
 use chrono::Local;
 
-use crate::registry::PUT_REGISTRY;
+use crate::put_registry::PUT_REGISTRY;
 
 pub fn format_title(title: Option<&str>, index: Option<usize>) -> String {
     let date = Local::now().format("%Y-%m-%d-%H%M%S");
