+++ tlspuffin/src/put_registry.rs
@@ -1,6 +1,6 @@
 use crate::{
     agent::PutName,
-    concretize::{Config, Put},
+    put::{Config, Put},
 };
 
 pub struct PutRegistry<const N: usize>([fn() -> Box<dyn Factory>; N]);
